Revert "Refactor initial validation"
This reverts commit 9d384c79bf11a3e92de6bd81414616b9d7561c4a.
