+++ gpMgmt/bin/gprebalance_modules/Makefile
@@ -3,7 +3,7 @@
 top_builddir = ../../..
 include $(top_builddir)/src/Makefile.global
 
-PROGRAMS= __init__.py rebalance_validator.py rebalance.py models.py cli.py
+PROGRAMS= __init__.py rebalance_validator.py rebalance.py
 
 installdirs:
 	$(MKDIR_P) '$(DESTDIR)$(bindir)/gprebalance_modules'
